+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Added `flake-utils-plus` JUST to generate the `registry` and `/etc/nix/inputs/<input>`
 - Changed `nixpkgs` in the `registry` to refer to the `nixpkgs` used when building the system
 - Make `dotfiles` version accessible through `nixos-version --json` to allow `polybar` to check the currently running `dotfiles`
 - Changed `fonts` into `hmModule`